# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShopFlow Platform API
  description: |
    REST API for the wholesale/retail shop management platform (React web and React Native clients).

    * Authentication: mobile number + OTP → short-lived bearer access token (JWT) plus a rotating refresh token.
    * Authorization: role (OWNER, ADMIN, CUSTOMER) plus permission codes; customers can only access their own data.
    * Envelope: every JSON response is `{ success, data, message, pagination?, error?, requestId }`.
    * Errors: `error.code` is one of the documented business codes (e.g. INSUFFICIENT_STOCK, CREDIT_LIMIT_EXCEEDED).
    * Money: decimal numbers with 2 places, calculated only by the backend.
    * Lists: `page` (1-based), `pageSize` (max 100), `sort=field,asc|desc`.
    * Retries: send `Idempotency-Key` on order creation, payment recording, invoice generation and WhatsApp sends.
  version: v1
servers:
- url: /
  description: Current host
...
paths:
...
  /api/v1/customer-registration:
    post:
      tags:
      - Customer Registration
      summary: Submit a customer registration
      description: "Requires the registration token returned by /auth/otp/verify as\
        \ the bearer token. Creates the account with status PENDING_APPROVAL and signs\
        \ the customer in with a limited session. Errors: VALIDATION_ERROR, CUSTOMER_ALREADY_REGISTERED."
      operationId: register
      parameters:
      - name: X-Client-Type
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAuthResponse"
      security:
      - bearerAuth: []
...
  /api/v1/dashboard/customer:
    get:
      tags:
      - Dashboard
      summary: Customer home summary
      description: "Outstanding, open orders, recent orders and recently ordered products\
        \ for the signed-in customer."
      operationId: customer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
      security:
      - bearerAuth: []
...
  /api/v1/customer-registration/status:
    get:
      tags:
      - Customer Registration
      summary: Get my registration status
      description: "PENDING_APPROVAL, APPROVED, REJECTED or BLOCKED."
      operationId: status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRegistrationStatusResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldIssue"
...
    FieldIssue:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    AddressRequest:
      type: object
      properties:
        label:
          type: string
          maxLength: 60
          minLength: 0
        addressLine1:
          type: string
          maxLength: 200
          minLength: 0
        addressLine2:
          type: string
          maxLength: 200
          minLength: 0
        city:
          type: string
          maxLength: 100
          minLength: 0
        state:
          type: string
          maxLength: 100
          minLength: 0
        stateCode:
          type: string
          pattern: "^[0-9]{2}$"
        pincode:
          type: string
          minLength: 1
          pattern: "^[1-9][0-9]{5}$"
        isDefault:
          type: boolean
      required:
      - addressLine1
      - city
      - pincode
      - state
...
    RegistrationRequest:
      type: object
      properties:
        shopName:
          type: string
          maxLength: 200
          minLength: 0
        contactName:
          type: string
          maxLength: 200
          minLength: 0
        address:
          $ref: "#/components/schemas/AddressRequest"
        gstin:
          type: string
          pattern: "^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z][1-9A-Z]Z[0-9A-Z]$"
        pan:
          type: string
          pattern: "^[A-Z]{5}[0-9]{4}[A-Z]$"
        email:
          type: string
          format: email
          maxLength: 200
          minLength: 0
        alternateMobile:
          type: string
      required:
      - address
      - contactName
      - shopName
    ApiResponseAuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AuthResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    AuthResponse:
      type: object
      properties:
        registrationRequired:
          type: boolean
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        refreshToken:
          type: string
        refreshTokenExpiresAt:
          type: string
          format: date-time
        registrationToken:
          type: string
        registrationTokenExpiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/MeResponse"
    CustomerInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerCode:
          type: string
        shopName:
          type: string
        status:
          type: string
    MeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        customer:
          $ref: "#/components/schemas/CustomerInfo"
...
    ApiResponseMapStringObject:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: {}
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
    ApiResponseRegistrationStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/RegistrationStatusResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    RegistrationStatusResponse:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        customerCode:
          type: string
        status:
          type: string
        reason:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT